shoreditch.php - Only load assets if Shoreditch is active
This follows up on the issues from:

https://github.com/civicrm/org.civicrm.shoreditch/pull/399#issuecomment-516005533

Let's recount a few things:

(1) The core theme mechanism is *mainly* a way to swap a CSS file. It has one or two fancy tricks
for how it does that (`url_callback`), but basically it's about swapping a CSS file.

(2) As it exists today, Shoreditch (and most likely some of the extensions tuned to Shoreditch) have
these expecations:

* (2a) Instead of swapping `civicrm.css`, it supplements it (i.e. appending
       `custom-civicrm.css` further down the list of CSS files).
* (2b) `bootstrap.css` should be activated on a wide range of pages
* (2c) Various JS helpers should be activated on all pages

These behaviors go beyond what's spec'd in core's theme layer. (Granted, 2a is sorta
addressed with `url_callback` - but 2b and 2c are currently outside the scope.)

(3) The main source of conflict when folks deploy Shoreditch has been in the
frontend -- e.g.  you have a WordPress theme with Bootstrap{CSS,JS}; if you
use Shoreditch on the Civi backend, then it can interfere on the WP
frontend.

With this patch, Shoreditch continues addressing expectations (2a,2b,2c) in
the same way as it did before.  However, *it only does these things if
Shoreditch is active on the current pageview.* Thus, if you have a WordPress
site with `theme_backend=shoreditch` and `theme_front=default`, then
Shoreditch won't interefere on the frontend.

This still leaves some edge-cases unaddressed (e.g.  *using Shoreditch for
frontend pages on a WP with Bootstrap*).  Addressing those edge-cases will
be more complicated (e.g.  talking about 2b,2c more exhaustively).

However, we're under time-pressure, and this approach should address a major
pain-point in a simple fashion.