Fix missing slide buttons in tables.
In the current latest version of CiviCRM (5.27.0) we noticed the slide buttons where missing inside the manage contribution pages and manage events tables. This pull request fixes this issue and also makes the label visible again.
I find it much more usable when the labels are visible. A user previously had to guess which options are underneath the bullets. This makes the available actions much clearer, also for less experienced users.

Any feedback is welcome!